Move sendbird to it's own controller
