+++ src/controllers/auth.js
@@ -13,70 +13,13 @@ const appleSigninAuth = require('apple-signin-auth');
 const LeadUsers = require('../models/lead_users');
 const { updatePurposes } = require('./interests');
 const { sendWelcomeEmail, sendVerificationEmail } = require('./email');
+const sendbird = require('./sendbird');
 
 const sesClient = new SESClient({ region: process.env.AWS_REGION });
 function generateToken(session_id, user_id) {
 	return jwt.sign({ session_id, user_id }, process.env.JWT_SECRET, {
 		expiresIn: '30d',
 	});
-}
-async function createSendbirdUser(userId) {
-	const user = await usersController.getUser(userId);
-
-	const sendbirdUserId = `${process.env.NODE_ENV}_${userId}`;
-
-	const response = await fetch(
-		`https://api-${process.env.SENDBIRD_APP_ID}.sendbird.com/v3/users`,
-		{
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-				'Api-Token': process.env.SENDBIRD_TOKEN,
-			},
-			body: JSON.stringify({
-				user_id: sendbirdUserId,
-				nickname: user.full_name,
-				profile_url: user.avatar,
-			}),
-		}
-	);
-
-	if (!response.ok) {
-		const errorData = await response.json();
-
-		throw new Error(errorData);
-	}
-
-	return generateSendbirdToken(userId);
-}
-
-async function generateSendbirdToken(userId) {
-	const response = await fetch(
-		`https://api-${process.env.SENDBIRD_APP_ID}.sendbird.com/v3/users/${process.env.NODE_ENV}_${userId}/token`,
-		{
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-				'Api-Token': process.env.SENDBIRD_TOKEN,
-			},
-			body: JSON.stringify({}),
-		}
-	);
-
-	if (!response.ok) {
-		const errorData = await response.json();
-		console.log(errorData);
-
-		if (errorData.code === 400201) {
-			return createSendbirdUser(userId);
-		}
-		console.log(`Failed to generate token: ${errorData.message}`);
-		throw new Error(errorData);
-	}
-
-	const data = await response.json();
-	console.log(`Access token generated for user ${userId}: ${data.token}`);
-	return data.token;
 }
 
 async function updateUserFromLead(email) {
@@ -303,7 +246,9 @@ async function verifyEmail({
 		});
 
 		const token = generateToken(newSession.id, user.id);
-		const sendbirdToken = await generateSendbirdToken(user.id);
+		const sendbirdToken = await sendbird.generateSendbirdToken(
+			user.id
+		);
 		const userInfo = await usersController.getUser(user.id);
 
 		return Promise.resolve({
@@ -372,7 +317,9 @@ async function login({ email, password, session_token, device_id, platform }) {
 		}
 
 		const userInfo = await usersController.getUser(user.id);
-		const sendbirdToken = await generateSendbirdToken(user.id);
+		const sendbirdToken = await sendbird.generateSendbirdToken(
+			user.id
+		);
 
 		return Promise.resolve({
 			data: {
@@ -437,7 +384,9 @@ async function googleLogin({ token, session_token, device_id, platform }) {
 		});
 
 		const access_token = generateToken(newSession.id, user.id);
-		const sendbirdToken = await generateSendbirdToken(user.id);
+		const sendbirdToken = await sendbird.generateSendbirdToken(
+			user.id
+		);
 
 		if (!user.is_active) {
 			await Users.update(
@@ -527,9 +476,8 @@ async function appleLogin({
 				newSession.id,
 				user.id
 			);
-			const sendbirdToken = await generateSendbirdToken(
-				user.id
-			);
+			const sendbirdToken =
+				await sendbird.generateSendbirdToken(user.id);
 
 			if (!user.is_active) {
 				await Users.update(
@@ -617,5 +565,4 @@ module.exports = {
 	logout,
 	updatePassword,
 	resendVerification,
-	generateSendbirdToken,
 };
